♻️ refactor(cli): move run/result/evidence/report/init under `lh acceptance`
Reorganize the CLI so the acceptance decision aggregate is the home for the
whole review loop, and `verify` keeps only the checker machinery (criteria,
rubrics, per-run plans).

- New canonical tree: `lh acceptance init` and `lh acceptance run …` (ingest,
  list, get, delete, result{ingest,list,submit,decision}, evidence, report).
  `ingest-report` becomes `acceptance run ingest`; a run is one immutable round.
- `verify` retains criterion / rubric / plan / execute / install / submit-less
  machinery. Everything moved keeps a deprecated `lh verify …` alias (same flags,
  same action, one deprecation warning) for a few releases.
- Extract shared helpers into verifyHelpers.ts (re-exported from ./verify for
  existing importers); action bodies live once in acceptanceRun.ts and wire both
  the canonical and deprecated spellings, so they can't drift.
- `init` now defaults to the `acceptance` skill bundle.

Tests: 52 passing — deprecated aliases still dispatch, the canonical run tree
dispatches, and the run subtree is absent from the deprecated `verify acceptance`
alias.
