+++ apps/cli/src/commands/verify.ts
@@ -1,593 +1,44 @@
-import { execFileSync } from 'node:child_process';
-import { existsSync, mkdirSync, readFileSync, writeFileSync } from 'node:fs';
-import path from 'node:path';
-
-import type {
-  AcceptanceSubjectType,
-  VerifyRunOrigin,
-  VerifyRunScenario,
-  VerifySurface,
-} from '@lobechat/const/verify';
-import {
-  acceptanceSubjectTypes,
-  normalizeVerifySurface,
-  verifyEvidenceTypes,
-  verifyRunScenarios,
-  verifySurfaces,
-} from '@lobechat/const/verify';
 import type { Command } from 'commander';
 import pc from 'picocolors';
 
 import { getTrpcClient } from '../api/client';
 import { confirm, outputJson, printTable, timeAgo, truncate } from '../utils/format';
 import { log } from '../utils/logger';
-import { uploadLocalFile } from '../utils/uploadLocalFile';
+import { attachDeprecatedVerifyRunAliases } from './acceptanceRun';
 import { registerAcceptanceCommands } from './verifyAcceptance';
+import {
+  assertEnum,
+  ON_FAIL,
+  type OnFail,
+  parseConfig,
+  printResults,
+  VERIFIER_TYPES,
+  type VerifierType,
+} from './verifyHelpers';
 import { registerVerifyInstallCommand } from './verifyInstall';
 
-// ── Helpers ────────────────────────────────────────────────
-
-type VerifierType = 'agent' | 'llm' | 'program';
-type OnFail = 'auto_repair' | 'manual';
-type Decision = 'accepted' | 'overridden' | 'rejected';
-
-const VERIFIER_TYPES: VerifierType[] = ['program', 'agent', 'llm'];
-const ON_FAIL: OnFail[] = ['manual', 'auto_repair'];
-const DECISIONS: Decision[] = ['accepted', 'rejected', 'overridden'];
-/** The evidence media a plan item may require — the same closed set the executor gates on. */
-const EVIDENCE_TYPES = verifyEvidenceTypes;
-
-function parseConfig(raw?: string): Record<string, unknown> | undefined {
-  if (!raw) return undefined;
-  try {
-    return JSON.parse(raw);
-  } catch {
-    log.error('--config must be valid JSON');
-    process.exit(1);
-  }
-}
-
-function assertEnum<T extends string>(value: T | undefined, allowed: T[], flag: string): void {
-  if (value !== undefined && !allowed.includes(value)) {
-    log.error(`${flag} must be one of: ${allowed.join(', ')}`);
-    process.exit(1);
-  }
-}
-
-type Verdict = 'failed' | 'passed' | 'uncertain';
-type EvidenceType = 'dom_snapshot' | 'gif' | 'screenshot' | 'text' | 'transcript' | 'video';
-
-const INLINE_TEXT_EVIDENCE_LIMIT = 5000;
-const INLINE_TEXT_EVIDENCE_TYPES = new Set<EvidenceType>(['dom_snapshot', 'text', 'transcript']);
-
-/** Map a free-form case/summary result token onto the verify verdict vocabulary. */
-function toVerdict(raw: unknown): Verdict {
-  const s = String(raw ?? '').toLowerCase();
-  if (['pass', 'passed', 'ok', 'success'].includes(s)) return 'passed';
-  if (['fail', 'failed', 'error'].includes(s)) return 'failed';
-  return 'uncertain'; // partial / blocked / skipped / pending / unknown
-}
-
-/**
- * The report's headline verdict when the author didn't set `summary.verdict`:
- * derived from the ingested cases, so a report can never ship verdict-less and
- * render as a permanent "?" in every list surface.
- */
-export function deriveReportVerdict(cases: unknown[]): Verdict | undefined {
-  const verdicts = cases.map((c) => toVerdict((c as any)?.result ?? (c as any)?.verdict));
-  if (verdicts.length === 0) return undefined;
-  if (verdicts.includes('failed')) return 'failed';
-  if (verdicts.includes('uncertain')) return 'uncertain';
-  return 'passed';
-}
-
-/** Pick an evidence medium from a file extension. */
-function evidenceTypeForFile(file: string): EvidenceType {
-  const ext = path.extname(file).toLowerCase().slice(1);
-  if (ext === 'gif') return 'gif';
-  if (['png', 'jpg', 'jpeg', 'webp', 'svg', 'bmp'].includes(ext)) return 'screenshot';
-  if (['mp4', 'webm', 'mov', 'm4v'].includes(ext)) return 'video';
-  if (['html', 'htm'].includes(ext)) return 'dom_snapshot';
-  return 'text';
-}
-
-function inlineTextEvidenceForFile(file: string, type: EvidenceType | string): string | undefined {
-  if (!INLINE_TEXT_EVIDENCE_TYPES.has(type as EvidenceType)) return undefined;
-
-  try {
-    const buffer = readFileSync(file);
-    if (buffer.includes(0)) return undefined;
-
-    const content = buffer.toString('utf8');
-    return content.length > 0 && content.length < INLINE_TEXT_EVIDENCE_LIMIT ? content : undefined;
-  } catch {
-    return undefined;
-  }
-}
-
-/** Normalize a case's `evidence` field (string | string[] | {path}[]) to path strings. */
-interface ReportEvidenceInput {
-  comparison?: {
-    id: string;
-    label?: string;
-    layout?: 'horizontal' | 'vertical';
-    role: 'after' | 'before';
-  };
-  description?: string;
-  path: string;
-}
-
-export function reportEvidence(evidence: unknown): ReportEvidenceInput[] {
-  if (!evidence) return [];
-  const arr = Array.isArray(evidence) ? evidence : [evidence];
-  return arr
-    .map((e): ReportEvidenceInput | null => {
-      if (typeof e === 'string') return { path: e };
-      const value = objectValue(e);
-      const evidencePath = value && firstString(value.path, value.file);
-      if (!evidencePath) return null;
-      const comparison = objectValue(value.comparison);
-      const role = comparison?.role;
-      const id = firstString(comparison?.id);
-      // The report viewer pairs on `id` and drops any comparison lacking one, so
-      // an id-less half could never render side by side. Warn rather than upload
-      // a comparison that is silently downgraded to an ordinary image.
-      if (comparison && !(id && (role === 'before' || role === 'after'))) {
-        log.warn(
-          `evidence ${evidencePath}: comparison needs both a string "id" and role "before"/"after" — ignoring it`,
-        );
-      }
-      const layout = comparison?.layout === 'vertical' ? 'vertical' : undefined;
-      return {
-        comparison:
-          id && (role === 'before' || role === 'after')
-            ? { id, label: firstString(comparison?.label), layout, role }
-            : undefined,
-        description: firstString(value.description, value.desc),
-        path: evidencePath,
-      };
-    })
-    .filter((item): item is ReportEvidenceInput => item !== null);
-}
-
-function firstString(...values: unknown[]): string | undefined {
-  return values.find((v): v is string => typeof v === 'string' && v.length > 0);
-}
-
-function firstStringOrNumber(...values: unknown[]): string | number | undefined {
-  return values.find(
-    (v): v is string | number => (typeof v === 'string' && v.length > 0) || typeof v === 'number',
-  );
-}
-
-function objectValue(value: unknown): Record<string, unknown> | undefined {
-  return value && typeof value === 'object' && !Array.isArray(value)
-    ? (value as Record<string, unknown>)
-    : undefined;
-}
-
-function safeWebUrl(value: string | undefined): string | undefined {
-  if (!value) return undefined;
-
-  try {
-    const url = new URL(value);
-    return url.protocol === 'http:' || url.protocol === 'https:' ? url.toString() : undefined;
-  } catch {
-    return undefined;
-  }
-}
-
-/** Normalize common agent-testing PR shapes into the verify coding scope. */
-function pullRequestFromResult(result: Record<string, unknown>) {
-  const pr = objectValue(result.pullRequest) ?? objectValue(result.pr);
-  const url = safeWebUrl(
-    firstString(
-      pr?.url,
-      pr?.htmlUrl,
-      pr?.html_url,
-      result.pullRequestUrl,
-      result.prUrl,
-      typeof result.pr === 'string' ? result.pr : undefined,
-    ),
-  );
-  const number = firstStringOrNumber(pr?.number, result.pullRequestNumber, result.prNumber);
-  const title = firstString(pr?.title, result.pullRequestTitle, result.prTitle);
-  const entries = Object.entries({ number, title, url }).filter(([, v]) => v !== undefined);
-
-  return entries.length > 0 ? Object.fromEntries(entries) : undefined;
-}
-
-/**
- * The PR for a branch, asked of `gh`. A report almost always verifies a branch
- * that already has one, but the author has to remember to write it down — and
- * mostly doesn't, so the report loses its single most useful outbound link.
- * Best-effort by design: no `gh`, not a repo, not authenticated, or no PR for
- * the branch all mean "no PR", never a failed publish.
- */
-function pullRequestFromBranch(branch: string | undefined) {
-  if (!branch) return undefined;
-
-  try {
-    const raw = execFileSync(
-      'gh',
-      ['pr', 'view', branch, '--json', 'number,title,url'],
-      // `gh` writes its "no pull requests found" diagnostics to stderr; keep them
-      // off the CLI's own output.
-      { encoding: 'utf8', stdio: ['ignore', 'pipe', 'ignore'], timeout: 10_000 },
-    );
-    const parsed = objectValue(JSON.parse(raw));
-    const url = safeWebUrl(firstString(parsed?.url));
-    const number = firstStringOrNumber(parsed?.number);
-    const title = firstString(parsed?.title);
-    const entries = Object.entries({ number, title, url }).filter(([, v]) => v !== undefined);
-
-    return entries.length > 0 ? Object.fromEntries(entries) : undefined;
-  } catch {
-    return undefined;
-  }
-}
-
-/**
- * The surfaces the report says it exercised, as canonical values.
- *
- * Strict on purpose: an unrecognized surface is a hard error, not a silently
- * dropped value. Free-form surfaces are how the field rotted — prose, runtime
- * modes ("packaged build") and test kinds ("unit", "type-check") all ended up in
- * it, and none of them render. Failing here puts the fix in the author's hands
- * while they still have the context to make it.
- */
-export function surfacesFromResult(result: Record<string, unknown>): VerifySurface[] | undefined {
-  if (!Array.isArray(result.surfaces)) return undefined;
-
-  const raw = result.surfaces.filter((s: unknown): s is string => typeof s === 'string');
-  const canonical: VerifySurface[] = [];
-  const rejected: string[] = [];
-
-  for (const value of raw) {
-    const surface = normalizeVerifySurface(value);
-    if (surface) {
-      if (!canonical.includes(surface)) canonical.push(surface);
-    } else {
-      rejected.push(value);
-    }
-  }
-
-  if (rejected.length > 0) {
-    log.error(
-      `result.json "surfaces" must name the product surface a check ran on, one of: ${verifySurfaces.join(', ')}`,
-    );
-    log.error(`  rejected: ${rejected.map((v) => JSON.stringify(v)).join(', ')}`);
-    log.error(
-      '  Runtime detail ("packaged build", "CDP dev instance") belongs on a plan item\'s "method"; a test kind ("unit", "backend") is not a surface — a backend change verified through the CLI has surface "cli".',
-    );
-    process.exit(1);
-  }
-
-  return canonical.length > 0 ? canonical : undefined;
-}
-
-/**
- * What kind of delivery the report verified. Defaults to `coding` — the
- * agent-testing harness's home turf — and rejects an unknown value rather than
- * storing a scenario nothing renders (mirrors the strict surface policy: the
- * author still has the context to fix it).
- */
-export function scenarioFromResult(result: Record<string, unknown>): VerifyRunScenario {
-  const raw = result.scenario;
-  if (raw === undefined) return 'coding';
-  if (typeof raw === 'string' && (verifyRunScenarios as readonly string[]).includes(raw)) {
-    return raw as VerifyRunScenario;
-  }
-
-  log.error(
-    `result.json "scenario" must be one of: ${verifyRunScenarios.join(', ')} — rejected: ${JSON.stringify(raw)}`,
-  );
-  process.exit(1);
-}
-
-/**
- * A non-coding scenario's scope, passed through from result.json `context` as
- * the scenario's own bag (the server stores it as-is; known shapes live in
- * `@lobechat/types`). Top-level `entry` / `createdAt` are lifted as defaults so
- * every scenario gets the shared provenance fields for free; explicit `context`
- * keys win.
- */
-export function genericContextFromResult(
-  result: Record<string, unknown>,
-): Record<string, unknown> | undefined {
-  const bag = objectValue(result.context) ?? {};
-  const entries = Object.entries({
-    entry: firstString(result.entry),
-    testedAt: firstString(result.createdAt),
-    ...bag,
-  }).filter(([, v]) => v !== undefined);
-
-  return entries.length > 0 ? Object.fromEntries(entries) : undefined;
-}
-
-/** Reject an out-of-vocabulary plan value rather than storing a word nothing reads. */
-function assertPlanEnum<T extends string>(
-  value: string,
-  allowed: readonly T[],
-  field: string,
-  itemId: string,
-): T {
-  if ((allowed as readonly string[]).includes(value)) return value as T;
-
-  log.error(`result.json plan item "${itemId}": "${field}" must be one of: ${allowed.join(', ')}`);
-  log.error(`  rejected: ${JSON.stringify(value)}`);
-  process.exit(1);
-}
-
-/**
- * The checks the run set out to make, as frozen plan items.
- *
- * Two of the fields are a closed vocabulary, not free text, because the pipeline
- * actually acts on them:
- *
- * - `verifier` → {@link VerifyCheckItem.verifierType} (`program | agent | llm`):
- *   how the item is judged. Defaults to `agent`, but a command-asserted check is
- *   `program` and mislabelling it hides what actually produced the verdict.
- * - `requiredEvidence` → `verifierConfig.requiredEvidence`
- *   ({@link RequiredEvidenceSpec}, medium from {@link verifyEvidenceTypes}):
- *   the artifact the item MUST produce. The executor's coverage gate fails an
- *   item whose required medium is missing, so this is enforcement, not a label —
- *   an unknown medium would silently gate on nothing.
- *
- * `method` / `expected` stay prose: they are the author's intent in words, which
- * no enum can carry. Everything else the frozen {@link VerifyCheckItem} needs is
- * filled in here, so an author writes only `{ id, title }` plus what they mean.
- *
- * Returns `undefined` when the report declares no `plan` field. A `plan` that is
- * present but empty returns `[]`, recording an explicitly empty plan in this
- * immutable snapshot.
- */
-export function planFromResult(result: Record<string, unknown>) {
-  if (!Array.isArray(result.plan)) return undefined;
-
-  const items = result.plan.flatMap((entry: unknown, index: number) => {
-    const item = objectValue(entry);
-    const title = firstString(item?.title, item?.name);
-    // An item with no title names no check — it can't be rendered or paired.
-    if (!item || !title) return [];
-
-    const id = String(item.id ?? `case-${index + 1}`);
-    const method = firstString(item.method, item.how);
-    const expected = firstString(item.expected, item.expectation);
-
-    const verifierRaw = firstString(item.verifier, item.verifierType);
-    const verifierType = verifierRaw
-      ? assertPlanEnum(verifierRaw, VERIFIER_TYPES, 'verifier', id)
-      : ('agent' as const);
-
-    const requiredEvidence = Array.isArray(item.requiredEvidence)
-      ? item.requiredEvidence.flatMap((spec: unknown) => {
-          // Accept the bare medium (`"screenshot"`) or the full spec object.
-          const record = objectValue(spec);
-          const raw = typeof spec === 'string' ? spec : firstString(record?.type);
-          if (!raw) return [];
-
-          return [
-            {
-              hint: firstString(record?.hint),
-              type: assertPlanEnum(raw, EVIDENCE_TYPES, 'requiredEvidence', id),
-            },
-          ];
-        })
-      : [];
-
-    // Per-item surface: the acceptance union groups checks by it. Normalized to
-    // the closed set (electron → desktop); an unknown value is dropped loudly
-    // rather than stored as a mystery chip.
-    const surfaceRaw = firstString(item.surface);
-    const surface = surfaceRaw ? normalizeVerifySurface(surfaceRaw) : null;
-    if (surfaceRaw && !surface) {
-      log.warn(
-        `plan item "${id}": surface "${surfaceRaw}" names no product surface (expected ${verifySurfaces.join('/')}) — dropping it`,
-      );
-    }
-
-    // The acceptance union groups by category and folds superseded ids into
-    // the new item's iteration timeline — both authored by the harness.
-    const category = firstString(item.category, item.group);
-    const supersedes = Array.isArray(item.supersedes)
-      ? item.supersedes.filter((value: unknown): value is string => typeof value === 'string')
-      : [];
-
-    return [
-      {
-        ...(category === undefined ? {} : { category }),
-        description: firstString(item.description),
-        id,
-        index,
-        onFail: 'manual' as const,
-        required: typeof item.required === 'boolean' ? item.required : true,
-        ...(supersedes.length > 0 ? { supersedes } : {}),
-        title,
-        verifierConfig: {
-          ...(method === undefined ? {} : { method }),
-          ...(expected === undefined ? {} : { expected }),
-          ...(requiredEvidence.length > 0 ? { requiredEvidence } : {}),
-          ...(surface ? { surface } : {}),
-        },
-        verifierType,
-      },
-    ];
-  });
-
-  // `[]` is meaningful — it clears a stale plan. Only an absent `plan` field
-  // (handled above) means "don't touch what's stored".
-  return items;
-}
-
-/**
- * The LobeHub conversation this harness is running inside, read off the env the
- * agent runtime echoes into the child process. Lets a report published from an
- * in-app agent link back to the session that produced it with no flags to
- * remember. Absent (a plain terminal) → no origin, which is not an error.
- *
- * Reads the env and ONLY the env. `--operation` is the opposite relation — the
- * Agent Run this session *verifies* — so letting it fall through to here would
- * attribute the report to the run under test instead of the run that wrote it,
- * corrupting the very provenance this records. The two ids are independent and
- * may legitimately differ in the same publish.
- */
-export function originFromEnv(): VerifyRunOrigin | undefined {
-  const origin: VerifyRunOrigin = {
-    agentId: firstString(process.env.LOBEHUB_AGENT_ID),
-    operationId: firstString(process.env.LOBEHUB_OPERATION_ID),
-    topicId: firstString(process.env.LOBEHUB_TOPIC_ID),
-  };
-
-  return Object.values(origin).some(Boolean) ? origin : undefined;
-}
-
-function metadataForReport(
-  result: Record<string, unknown>,
-  existingMetadata?: unknown,
-  origin?: VerifyRunOrigin,
-): Record<string, unknown> | undefined {
-  const hasInteractionCost = Object.prototype.hasOwnProperty.call(result, 'interactionCost');
-  // Nothing to write is not the same as writing an empty metadata bag.
-  if (!hasInteractionCost && !origin) return undefined;
-
-  const metadata = { ...objectValue(existingMetadata) };
-
-  if (hasInteractionCost) {
-    const interactionCost = objectValue(result.interactionCost);
-
-    if (interactionCost) {
-      metadata.interactionCost = interactionCost;
-    } else {
-      delete metadata.interactionCost;
-    }
-  }
-
-  if (origin) metadata.origin = origin;
-
-  return metadata;
-}
-
-/** A parsed acceptance subject reference (`task:<id>` / `topic:<id>` / `document:<id>`). */
-export interface AcceptanceSubjectRef {
-  subjectId: string;
-  subjectType: AcceptanceSubjectType;
-}
-
-/**
- * Parse a `type:id` subject reference. Returns null on anything malformed —
- * callers decide whether that is an error (an explicit `--subject`) or a
- * silently absent field (result.json).
- */
-export function parseSubjectRef(raw: unknown): AcceptanceSubjectRef | null {
-  if (typeof raw !== 'string') return null;
-  const idx = raw.indexOf(':');
-  if (idx <= 0) return null;
-  const type = raw.slice(0, idx).trim().toLowerCase();
-  const id = raw.slice(idx + 1).trim();
-  if (!id || !(acceptanceSubjectTypes as readonly string[]).includes(type)) return null;
-  return { subjectId: id, subjectType: type as AcceptanceSubjectType };
-}
-
-/**
- * The acceptance subject a report attributes itself to, from result.json's
- * `subject` field — either `"task:<id>"` or `{ type, id, requirement? }`. An
- * explicit `--subject` flag wins over this.
- */
-export function subjectFromResult(result: Record<string, unknown>): {
-  ref: AcceptanceSubjectRef;
-  requirement?: string;
-} | null {
-  const raw = result.subject;
-  if (typeof raw === 'string') {
-    const ref = parseSubjectRef(raw);
-    return ref ? { ref } : null;
-  }
-  const value = objectValue(raw);
-  if (!value) return null;
-  const ref = parseSubjectRef(`${firstString(value.type) ?? ''}:${firstString(value.id) ?? ''}`);
-  return ref ? { ref, requirement: firstString(value.requirement) } : null;
-}
-
-/** Default acceptance subject for a report authored inside a LobeHub topic. */
-export function subjectFromEnv(): AcceptanceSubjectRef | null {
-  const topicId = firstString(process.env.LOBEHUB_TOPIC_ID);
-  return topicId ? { subjectId: topicId, subjectType: 'topic' } : null;
-}
+// Re-export the report/ingest helpers so existing importers keep resolving them
+// from './verify' (tests, and callers predating the verifyHelpers split).
+export * from './verifyHelpers';
 
 // ── Command Registration ───────────────────────────────────
 
 export function registerVerifyCommand(program: Command) {
   const verify = program
     .command('verify')
-    .description('Manage the Agent Run delivery checker (criteria, rubrics, plans, results)');
+    .description('Agent Run verification machinery — criteria, rubrics, and per-run check plans');
 
-  // `verify acceptance …` — subject-level acceptance aggregates.
-  // Legacy spelling — the canonical group is the first-class `lh acceptance`.
+  // `verify acceptance …` — legacy alias; the canonical group is the first-class
+  // `lh acceptance`.
   registerAcceptanceCommands(verify, { deprecated: true });
 
   // `verify install` — install the bundled agent-testing skill from this CLI
-  // package into the consumer repo. Distinct from `verify init` below, which
-  // pulls a named skill bundle live from the server.
+  // package into the consumer repo.
   registerVerifyInstallCommand(verify);
 
-  // ════════════ init (materialize the portable verify skill) ════════════
-  verify
-    .command('init')
-    .description('Write the portable verify skill into a working dir (.claude/skills/verify)')
-    .option('--dir <path>', 'Target working directory (default: current dir)')
-    .option('--skill <id>', 'Skill identifier to pull', 'verify')
-    .option('--force', 'Overwrite existing skill files')
-    .option('--json [fields]', 'Output JSON')
-    .action(
-      async (options: {
-        dir?: string;
-        force?: boolean;
-        json?: boolean | string;
-        skill: string;
-      }) => {
-        const client = await getTrpcClient();
-        // Pulled live from the server's deployed builtin-skills — always the latest.
-        const bundle = await client.verify.getSkillBundle.query({ identifier: options.skill });
-
-        const baseDir = options.dir ? path.resolve(options.dir) : process.cwd();
-        const skillDir = path.join(baseDir, '.claude', 'skills', bundle.identifier);
-
-        // path → content for SKILL.md plus every resource file.
-        const entries: [string, string][] = [
-          ['SKILL.md', bundle.content],
-          ...Object.entries(bundle.files),
-        ];
-
-        const written: string[] = [];
-        const skipped: string[] = [];
-        for (const [rel, content] of entries) {
-          const dest = path.join(skillDir, rel);
-          if (existsSync(dest) && !options.force) {
-            skipped.push(rel);
-            continue;
-          }
-          mkdirSync(path.dirname(dest), { recursive: true });
-          writeFileSync(dest, content, 'utf8');
-          written.push(rel);
-        }
-
-        const result = { dir: skillDir, skill: bundle.identifier, skipped, written };
-        if (options.json !== undefined) {
-          outputJson(result, typeof options.json === 'string' ? options.json : undefined);
-          return;
-        }
-        console.log(
-          `${pc.green('✓')} ${pc.bold(bundle.name)} skill → ${pc.dim(path.relative(process.cwd(), skillDir) || skillDir)}`,
-        );
-        console.log(
-          `  ${written.length} written${skipped.length ? `, ${skipped.length} skipped` : ''}`,
-        );
-        if (skipped.length > 0)
-          console.log(pc.dim(`  (skipped existing — pass --force to overwrite)`));
-      },
-    );
-
+  // Deprecated `lh verify …` spellings for the run/result/evidence/report/init
+  // commands now living under `lh acceptance`. Kept for a few releases.
+  attachDeprecatedVerifyRunAliases(verify);
   // ════════════ criteria ════════════
   const criterion = verify.command('criterion').description('Reusable pass/fail standards');
 
@@ -954,746 +405,4 @@ export function registerVerifyCommand(program: Command) {
         printResults(results);
       },
     );
-
-  // ════════════ run (verification session entity) ════════════
-  const run = verify.command('run').description('Verification sessions (verify_runs)');
-
-  run
-    .command('create')
-    .description('Create a standalone verification session')
-    .option('--source <source>', 'agent | agent-testing', 'agent-testing')
-    .option('--operation <id>', 'Link to an existing Agent Run')
-    .option('--title <title>', 'Session title')
-    .option('--goal <goal>', 'Goal/task being verified')
-    .option('--json [fields]', 'Output JSON')
-    .action(
-      async (options: {
-        goal?: string;
-        json?: boolean | string;
-        operation?: string;
-        source?: string;
-        title?: string;
-      }) => {
-        const client = await getTrpcClient();
-        const created = await client.verify.createRun.mutate({
-          goal: options.goal,
-          operationId: options.operation,
-          source: options.source as any,
-          title: options.title,
-        });
-        if (options.json !== undefined) {
-          outputJson(created, typeof options.json === 'string' ? options.json : undefined);
-          return;
-        }
-        console.log(`${pc.green('✓')} Created run ${pc.bold(created.id)}`);
-      },
-    );
-
-  run
-    .command('list')
-    .description('List recent verification sessions')
-    .option('--json [fields]', 'Output JSON')
-    .action(async (options: { json?: boolean | string }) => {
-      const client = await getTrpcClient();
-      const runs = await client.verify.listRuns.query();
-      if (options.json !== undefined) {
-        outputJson(runs, typeof options.json === 'string' ? options.json : undefined);
-        return;
-      }
-      if (runs.length === 0) return void console.log('No runs found.');
-      printTable(
-        runs.map((r: any) => [
-          r.id,
-          truncate(r.title || '', 40),
-          r.source,
-          r.status ?? '',
-          r.operationId ? 'agent' : 'standalone',
-          r.createdAt ? timeAgo(r.createdAt) : '',
-        ]),
-        ['ID', 'TITLE', 'SOURCE', 'STATUS', 'KIND', 'CREATED'],
-      );
-    });
-
-  run
-    .command('get <runId>')
-    .description('Show a verification session')
-    .option('--json [fields]', 'Output JSON')
-    .action(async (runId: string, options: { json?: boolean | string }) => {
-      const client = await getTrpcClient();
-      const item = await client.verify.getRun.query({ verifyRunId: runId });
-      if (options.json !== undefined) {
-        outputJson(item, typeof options.json === 'string' ? options.json : undefined);
-        return;
-      }
-      if (!item) return void console.log('Run not found.');
-      console.log(JSON.stringify(item, null, 2));
-    });
-
-  run
-    .command('delete <runId>')
-    .description('Delete a verification session (cascades its results, evidence and report)')
-    .option('-y, --yes', 'Skip the confirmation prompt')
-    .option('--json [fields]', 'Output JSON')
-    .action(async (runId: string, options: { json?: boolean | string; yes?: boolean }) => {
-      const client = await getTrpcClient();
-      if (!options.yes) {
-        const ok = await confirm(
-          `Delete run ${pc.bold(runId)} and all its results, evidence and report? This cannot be undone.`,
-        );
-        if (!ok) return void console.log('Aborted.');
-      }
-      const result = await client.verify.deleteRun.mutate({ verifyRunId: runId });
-      if (options.json !== undefined) {
-        outputJson(result, typeof options.json === 'string' ? options.json : undefined);
-        return;
-      }
-      console.log(`${pc.green('✓')} Deleted run ${pc.bold(result.id)}`);
-    });
-
-  // ════════════ result (check result entity) ════════════
-  const result = verify.command('result').description('Check results (verify_check_results)');
-
-  result
-    .command('ingest')
-    .description('Upsert one check result by (run, checkItemId) from a supplied verdict')
-    .requiredOption('--run <verifyRunId>', 'Target session id')
-    .requiredOption('--check <checkItemId>', 'Stable check item id within the session')
-    .requiredOption('--verdict <verdict>', 'passed|failed|uncertain')
-    .option('--title <title>', 'Check title')
-    .option('--index <n>', 'Display index')
-    .option('--confidence <n>', '0-1 confidence')
-    .option('--status <status>', 'pending|running|passed|failed|skipped (derived from verdict)')
-    .option('--evidence <text>', 'Key observation (stored as Toulmin evidence)')
-    .option('--suggestion <text>', 'Remediation hint')
-    .option('--soft', 'Non-blocking (required=false); defaults to blocking')
-    .option('--json [fields]', 'Output JSON')
-    .action(
-      async (options: {
-        check: string;
-        confidence?: string;
-        evidence?: string;
-        index?: string;
-        json?: boolean | string;
-        run: string;
-        soft?: boolean;
-        status?: string;
-        suggestion?: string;
-        title?: string;
-        verdict: string;
-      }) => {
-        const client = await getTrpcClient();
-        const created = await client.verify.ingestResult.mutate({
-          checkItemId: options.check,
-          checkItemIndex: options.index ? Number.parseInt(options.index, 10) : undefined,
-          checkItemTitle: options.title,
-          confidence: options.confidence ? Number.parseFloat(options.confidence) : undefined,
-          required: options.soft ? false : undefined,
-          status: options.status as any,
-          suggestion: options.suggestion,
-          toulmin: options.evidence ? { evidence: options.evidence } : undefined,
-          verdict: options.verdict as any,
-          verifyRunId: options.run,
-        });
-        if (options.json !== undefined) {
-          outputJson(created, typeof options.json === 'string' ? options.json : undefined);
-          return;
-        }
-        console.log(`${pc.green('✓')} Result ${pc.bold(created.id)} (${created.verdict})`);
-      },
-    );
-
-  result
-    .command('list')
-    .description('List check results — by session (--run) or by Agent Run (--operation)')
-    .option('--run <verifyRunId>', 'List by verification session')
-    .option('--operation <operationId>', 'List by Agent Run')
-    .option('--json [fields]', 'Output JSON')
-    .action(async (options: { json?: boolean | string; operation?: string; run?: string }) => {
-      if (!options.run && !options.operation) {
-        log.error('Provide either --run or --operation');
-        process.exit(1);
-      }
-      const client = await getTrpcClient();
-      const results = options.run
-        ? await client.verify.listResultsByRun.query({ verifyRunId: options.run })
-        : await client.verify.listResults.query({ operationId: options.operation! });
-      if (options.json !== undefined) {
-        outputJson(results, typeof options.json === 'string' ? options.json : undefined);
-        return;
-      }
-      if (results.length === 0) return void console.log('No results yet.');
-      printResults(results);
-    });
-
-  // ════════════ submit (builder self-evidence: result + evidence in one call) ════════════
-  verify
-    .command('submit')
-    .description('Submit a check item — upsert its result and attach evidence in one call')
-    .option('--run <verifyRunId>', 'Target verification session (or use --operation)')
-    .option('--operation <operationId>', 'Resolve the session from an Agent Run operation id')
-    .requiredOption('--item <checkItemId>', 'Plan item id (checkItemId)')
-    .option('--type <type>', 'screenshot|gif|video|text|dom_snapshot|transcript')
-    .option('--file <path>', 'Local file to upload as the evidence artifact')
-    .option('--content <text>', 'Inline text payload (instead of a file)')
-    .option('--verdict <verdict>', 'passed|failed|uncertain')
-    .option('--title <text>', 'Check item title snapshot')
-    .option('--by <capturedBy>', 'agent-browser|cdp|cli|program|llm_judge', 'cli')
-    .option('--desc <text>', 'Human-readable caption for the evidence')
-    .option('--json [fields]', 'Output JSON')
-    .action(
-      async (options: {
-        by?: string;
-        content?: string;
-        desc?: string;
-        file?: string;
-        item: string;
-        json?: boolean | string;
-        operation?: string;
-        run?: string;
-        title?: string;
-        type?: string;
-        verdict?: string;
-      }) => {
-        if (!options.run && !options.operation) {
-          log.error('Provide --run <verifyRunId> or --operation <operationId>');
-          process.exit(1);
-        }
-        const hasEvidence = Boolean(options.file) || Boolean(options.content);
-        if (Boolean(options.file) && Boolean(options.content)) {
-          log.error('Provide at most one of --file or --content');
-          process.exit(1);
-        }
-        if (hasEvidence && !options.type) {
-          log.error('--type is required when attaching evidence');
-          process.exit(1);
-        }
-        if (!hasEvidence && !options.verdict) {
-          log.error('Provide evidence (--file/--content) and/or a --verdict');
-          process.exit(1);
-        }
-        const client = await getTrpcClient();
-        let fileId: string | undefined;
-        let inlineContent = options.content;
-        if (options.file) {
-          inlineContent = inlineTextEvidenceForFile(options.file, options.type!);
-          if (inlineContent === undefined) {
-            const uploaded = await uploadLocalFile(client, options.file);
-            fileId = uploaded.id;
-          }
-        }
-        const evidence = hasEvidence
-          ? [
-              {
-                capturedBy: options.by as any,
-                content: inlineContent,
-                description: options.desc,
-                fileId,
-                type: options.type as any,
-              },
-            ]
-          : undefined;
-        const res = await client.verify.submitCheckEvidence.mutate({
-          checkItemId: options.item,
-          checkItemTitle: options.title,
-          evidence,
-          operationId: options.operation,
-          verdict: options.verdict as any,
-          verifyRunId: options.run,
-        });
-        if (options.json !== undefined) {
-          outputJson(res, typeof options.json === 'string' ? options.json : undefined);
-          return;
-        }
-        console.log(
-          `${pc.green('✓')} Submitted ${pc.bold(res.checkResult.id)}` +
-            `${res.checkResult.verdict ? ` (${res.checkResult.verdict})` : ''}` +
-            `${res.evidence.length > 0 ? ` +${res.evidence.length} evidence` : ''}`,
-        );
-      },
-    );
-
-  // ════════════ evidence (artifact entity) ════════════
-  const evidence = verify.command('evidence').description('Evidence artifacts (verify_evidence)');
-
-  evidence
-    .command('upload')
-    .description('Attach an evidence artifact (file or inline text) to a check result')
-    .requiredOption('--check <checkResultId>', 'Target check result id')
-    .requiredOption('--type <type>', 'screenshot|gif|video|text|dom_snapshot|transcript')
-    .option('--file <path>', 'Local file to upload as the artifact')
-    .option('--content <text>', 'Inline text payload (instead of a file)')
-    .option('--by <capturedBy>', 'agent-browser|cdp|cli|program|llm_judge', 'cli')
-    .option('--desc <text>', 'Human-readable caption')
-    .option('--json [fields]', 'Output JSON')
-    .action(
-      async (options: {
-        by?: string;
-        check: string;
-        content?: string;
-        desc?: string;
-        file?: string;
-        json?: boolean | string;
-        type: string;
-      }) => {
-        if (Boolean(options.file) === Boolean(options.content)) {
-          log.error('Provide exactly one of --file or --content');
-          process.exit(1);
-        }
-        const client = await getTrpcClient();
-        let fileId: string | undefined;
-        let inlineContent = options.content;
-        if (options.file) {
-          inlineContent = inlineTextEvidenceForFile(options.file, options.type);
-          if (inlineContent === undefined) {
-            const uploaded = await uploadLocalFile(client, options.file);
-            fileId = uploaded.id;
-          }
-        }
-        const ev = await client.verify.uploadEvidence.mutate({
-          capturedBy: options.by as any,
-          checkResultId: options.check,
-          content: inlineContent,
-          description: options.desc,
-          fileId,
-          type: options.type as any,
-        });
-        if (options.json !== undefined) {
-          outputJson(ev, typeof options.json === 'string' ? options.json : undefined);
-          return;
-        }
-        console.log(
-          `${pc.green('✓')} Evidence ${pc.bold(ev.id)}${fileId ? ` (file ${fileId})` : ''}`,
-        );
-      },
-    );
-
-  evidence
-    .command('list <checkResultId>')
-    .description('List evidence for a check result')
-    .option('--json [fields]', 'Output JSON')
-    .action(async (checkResultId: string, options: { json?: boolean | string }) => {
-      const client = await getTrpcClient();
-      const rows = await client.verify.listEvidence.query({ checkResultId });
-      if (options.json !== undefined) {
-        outputJson(rows, typeof options.json === 'string' ? options.json : undefined);
-        return;
-      }
-      if (rows.length === 0) return void console.log('No evidence.');
-      printTable(
-        rows.map((e: any) => [
-          e.id,
-          e.type,
-          e.capturedBy ?? '',
-          e.fileId ? 'file' : 'inline',
-          truncate(e.description || '', 40),
-        ]),
-        ['ID', 'TYPE', 'BY', 'PAYLOAD', 'DESC'],
-      );
-    });
-
-  evidence
-    .command('delete <evidenceId>')
-    .description('Delete an evidence artifact')
-    .option('--json [fields]', 'Output JSON')
-    .action(async (evidenceId: string, options: { json?: boolean | string }) => {
-      const client = await getTrpcClient();
-      const result = await client.verify.deleteEvidence.mutate({ id: evidenceId });
-      if (options.json !== undefined) {
-        outputJson(result, typeof options.json === 'string' ? options.json : undefined);
-        return;
-      }
-      console.log(`${pc.green('✓')} Deleted evidence ${pc.bold(result.id)}`);
-    });
-
-  // ════════════ report (narrative entity) ════════════
-  const report = verify.command('report').description('Verification reports (verify_reports)');
-
-  report
-    .command('upsert')
-    .description('Write (overwrite) the report for a session')
-    .requiredOption('--run <verifyRunId>', 'Target session id')
-    .option('--verdict <verdict>', 'passed|failed|uncertain')
-    .option('--summary <text>', 'Short summary')
-    .option('--content <markdown>', 'Full markdown body')
-    .option('--total <n>', 'Total checks')
-    .option('--passed <n>', 'Passed checks')
-    .option('--failed <n>', 'Failed checks')
-    .option('--uncertain <n>', 'Uncertain checks')
-    .option('--json [fields]', 'Output JSON')
-    .action(
-      async (options: {
-        content?: string;
-        failed?: string;
-        json?: boolean | string;
-        passed?: string;
-        run: string;
-        summary?: string;
-        total?: string;
-        uncertain?: string;
-        verdict?: string;
-      }) => {
-        const num = (s?: string) => (s === undefined ? undefined : Number.parseInt(s, 10));
-        const client = await getTrpcClient();
-        const created = await client.verify.upsertReport.mutate({
-          content: options.content,
-          failedChecks: num(options.failed),
-          passedChecks: num(options.passed),
-          summary: options.summary,
-          totalChecks: num(options.total),
-          uncertainChecks: num(options.uncertain),
-          verdict: options.verdict as any,
-          verifyRunId: options.run,
-        });
-        if (options.json !== undefined) {
-          outputJson(created, typeof options.json === 'string' ? options.json : undefined);
-          return;
-        }
-        console.log(`${pc.green('✓')} Report ${pc.bold(created.id)} (${created.verdict ?? '—'})`);
-      },
-    );
-
-  report
-    .command('get <runId>')
-    .description('Show the report for a session')
-    .option('--json [fields]', 'Output JSON')
-    .action(async (runId: string, options: { json?: boolean | string }) => {
-      const client = await getTrpcClient();
-      const item = await client.verify.getReport.query({ verifyRunId: runId });
-      if (options.json !== undefined) {
-        outputJson(item, typeof options.json === 'string' ? options.json : undefined);
-        return;
-      }
-      if (!item) return void console.log('No report.');
-      console.log(JSON.stringify(item, null, 2));
-    });
-
-  // ════════════ feedback ════════════
-  verify
-    .command('decision <resultId> <decision>')
-    .description(`Record human feedback on a result (${DECISIONS.join('|')})`)
-    .action(async (resultId: string, decision: Decision) => {
-      assertEnum(decision, DECISIONS, 'decision');
-      const client = await getTrpcClient();
-      await client.verify.submitDecision.mutate({ decision, resultId });
-      console.log(`${pc.green('✓')} Recorded ${pc.bold(decision)} on result ${pc.bold(resultId)}`);
-    });
-
-  // ════════════ ingest (aggregate convenience over the atomic commands) ════════════
-  verify
-    .command('ingest-report <reportDir>')
-    .description(
-      'Ingest a local agent-testing report (result.json + report.md + assets) as a verify session',
-    )
-    .option('--source <source>', 'agent | agent-testing', 'agent-testing')
-    .option('--operation <id>', 'Link the session to an existing Agent Run')
-    .option('--title <title>', 'Override the session title')
-    .option('--goal <goal>', 'The goal/task being verified')
-    .option(
-      '--subject <type:id>',
-      'Override the required acceptance subject (defaults to the current LOBEHUB_TOPIC_ID)',
-    )
-    .option(
-      '--requirement <text>',
-      'Acceptance requirement recorded when the aggregate is first created',
-    )
-    .option('--open', 'Print the in-app URL to open the report')
-    .option('--json [fields]', 'Output JSON')
-    .action(
-      async (
-        reportDir: string,
-        options: {
-          goal?: string;
-          json?: boolean | string;
-          open?: boolean;
-          operation?: string;
-          requirement?: string;
-          source?: string;
-          subject?: string;
-          title?: string;
-        },
-      ) => {
-        const dir = path.resolve(reportDir);
-        const resultPath = path.join(dir, 'result.json');
-        if (!existsSync(resultPath)) {
-          log.error(`result.json not found in ${dir}`);
-          process.exit(1);
-        }
-
-        let result: any;
-        try {
-          result = JSON.parse(readFileSync(resultPath, 'utf8'));
-        } catch {
-          log.error('result.json is not valid JSON');
-          process.exit(1);
-        }
-
-        const cases: any[] = Array.isArray(result.cases) ? result.cases : [];
-        const summary = result.summary ?? {};
-        const reportMdPath = path.join(dir, 'report.md');
-        const content = existsSync(reportMdPath) ? readFileSync(reportMdPath, 'utf8') : undefined;
-
-        // What kind of delivery this report verified (default: coding).
-        const scenario = scenarioFromResult(result);
-
-        // The scenario's context for the report's scope header. Coding lifts the
-        // well-known top-level fields (branch / commit / surfaces / PR); every
-        // other scenario passes result.json `context` through as its own bag.
-        // `pullRequest` is hoisted: the success output (text and --json) prints
-        // the PR link after the ingest, whatever the scenario resolved to.
-        let context: Record<string, unknown> | undefined;
-        let pullRequest: ReturnType<typeof pullRequestFromResult>;
-        if (scenario === 'coding') {
-          const branch = typeof result.branch === 'string' ? result.branch : undefined;
-          const surfaces = surfacesFromResult(result);
-          // An authored PR wins; otherwise ask `gh` what the branch's PR is, so the
-          // report links to it without the author having to remember the field.
-          pullRequest = pullRequestFromResult(result) ?? pullRequestFromBranch(branch);
-          const contextEntries = Object.entries({
-            branch,
-            commit: typeof result.commit === 'string' ? result.commit : undefined,
-            entry: typeof result.entry === 'string' ? result.entry : undefined,
-            pullRequest,
-            surfaces,
-            testedAt: typeof result.createdAt === 'string' ? result.createdAt : undefined,
-          }).filter(([, v]) => v !== undefined);
-          context = contextEntries.length > 0 ? Object.fromEntries(contextEntries) : undefined;
-        } else {
-          context = genericContextFromResult(result);
-        }
-
-        // What the run set out to check, written before it ran. Paired with the
-        // results by `id`, so the report can show a planned item that never ran.
-        const plan = planFromResult(result);
-
-        // Every agent-testing report belongs to an acceptance. Explicit CLI input
-        // wins, then result.json, then the authoring topic echoed by the runtime.
-        let subject = subjectFromResult(result);
-        if (options.subject) {
-          const ref = parseSubjectRef(options.subject);
-          if (!ref) {
-            log.error(
-              `--subject must be one of ${acceptanceSubjectTypes.map((t) => `${t}:<id>`).join(' | ')}`,
-            );
-            process.exit(1);
-          }
-          subject = { ref, requirement: subject?.requirement };
-        } else if (result.subject && !subject) {
-          log.error('result.json `subject` is malformed (expected "type:id" or {type,id})');
-          process.exit(1);
-        } else if (!subject) {
-          const ref = subjectFromEnv();
-          if (ref) subject = { ref };
-        }
-        if (!subject) {
-          log.error(
-            'Acceptance subject is required: run inside a LobeHub topic or pass --subject task:<id> | topic:<id> | document:<id>',
-          );
-          process.exit(1);
-        }
-        const requirement = options.requirement ?? subject?.requirement;
-
-        const client = await getTrpcClient();
-        const goal = options.goal ?? (typeof result.focus === 'string' ? result.focus : undefined);
-        const title = options.title ?? result.title;
-        // The title is the run's identity in every list surface — an untitled
-        // run renders as a placeholder forever, so say so before it ships.
-        if (!title) {
-          log.warn(
-            'result.json has no "title" — the run will list as untitled; set result.title (or pass --title)',
-          );
-        }
-        // The in-app conversation that ran this harness, if any (env-supplied).
-        // Strictly the authoring conversation. `--operation` names the Agent Run
-        // under test and is passed to `createRun` below — a different relation.
-        const origin = originFromEnv();
-        const newRunMetadata = metadataForReport(result, undefined, origin);
-
-        // Every ingest is a new immutable verification snapshot. A repair or
-        // re-verification is represented by another run on the same acceptance.
-        const run = await client.verify.createRun.mutate({
-          context,
-          goal,
-          metadata: newRunMetadata,
-          operationId: options.operation,
-          plan,
-          scenario,
-          source: options.source as any,
-          title,
-        });
-        const runId = run.id;
-
-        // 1c. Chain the session onto its subject's acceptance as the next round
-        //     BEFORE the report lands, so the report-time status rollup already
-        //     sees the aggregate.
-        const acceptance = await client.acceptance.ensure.mutate({
-          requirement,
-          subjectId: subject.ref.subjectId,
-          subjectType: subject.ref.subjectType,
-        });
-        const acceptanceId = acceptance.id;
-        await client.acceptance.attachRun.mutate({ acceptanceId, verifyRunId: runId });
-
-        // 2. Ingest each case as a check result + its evidence. `checkItemId` is
-        //    the stable key within this immutable run.
-        const seenCheckItemIds = new Set<string>();
-        let evidenceCount = 0;
-        let inlined = 0;
-        for (const [index, c] of cases.entries()) {
-          const checkItemId = String(c.id ?? c.checkItemId ?? `case-${index + 1}`);
-          seenCheckItemIds.add(checkItemId);
-          const verdict = toVerdict(c.result ?? c.status ?? c.verdict);
-          const observation = c.keyObservation ?? c.observation ?? c.note;
-          const checkResult = await client.verify.ingestResult.mutate({
-            checkItemId,
-            checkItemIndex: index,
-            checkItemTitle: c.name ?? c.case ?? c.title ?? checkItemId,
-            required: c.required ?? true,
-            // The case's key observation is recorded as Toulmin evidence; a real
-            // remediation hint (if the report provides one) goes to `suggestion`.
-            // Absent → explicit `null`, so this immutable snapshot records the
-            // absence instead of relying on upsert defaults.
-            suggestion: typeof c.suggestion === 'string' ? c.suggestion : null,
-            toulmin: typeof observation === 'string' ? { evidence: observation } : null,
-            verdict,
-            verifierType: 'agent',
-            verifyRunId: runId,
-          });
-
-          for (const evidenceInput of reportEvidence(c.evidence)) {
-            const rel = evidenceInput.path;
-            const abs = path.isAbsolute(rel) ? rel : path.join(dir, rel);
-            if (!existsSync(abs)) {
-              log.warn(`evidence not found, skipping: ${rel}`);
-              continue;
-            }
-            try {
-              const type = evidenceTypeForFile(abs);
-              const content = inlineTextEvidenceForFile(abs, type);
-              const file = content === undefined ? await uploadLocalFile(client, abs) : undefined;
-              await client.verify.uploadEvidence.mutate({
-                capturedBy: 'cli',
-                checkResultId: checkResult.id,
-                // The filename, not the case title — the title already heads the
-                // check card, so reusing it here just triples the same text.
-                content,
-                description: evidenceInput.description ?? path.basename(abs),
-                fileId: file?.id,
-                metadata: evidenceInput.comparison
-                  ? { comparison: evidenceInput.comparison }
-                  : undefined,
-                type,
-              });
-              evidenceCount += 1;
-              if (content !== undefined) inlined += 1;
-            } catch (e) {
-              // A stub/unreachable storage bucket (common in local dev) fails the
-              // file PUT — don't abort the whole ingest over one artifact; the
-              // session, results, and report are the deliverable.
-              log.warn(`evidence upload failed, skipping ${path.basename(abs)}: ${String(e)}`);
-            }
-          }
-        }
-
-        // 3. Write the report. `summary` is the overall conclusion (rendered at
-        //    the top of the report page); `content` is the full markdown detail.
-        const conclusion =
-          typeof summary.conclusion === 'string'
-            ? summary.conclusion
-            : typeof summary.note === 'string'
-              ? summary.note
-              : undefined;
-        // A 0-100 quality score lands on overallConfidence (0-1); the report page
-        // surfaces it as the `score` stat.
-        const score =
-          typeof summary.score === 'number'
-            ? Math.max(0, Math.min(1, summary.score / 100))
-            : undefined;
-        await client.verify.upsertReport.mutate({
-          content,
-          failedChecks: summary.failed,
-          overallConfidence: score,
-          passedChecks: summary.passed,
-          summary: conclusion,
-          totalChecks: summary.total ?? cases.length,
-          uncertainChecks: (summary.blocked ?? 0) + (summary.uncertain ?? 0) || undefined,
-          // An explicit summary.verdict wins; otherwise the headline is derived
-          // from the ingested cases (deriveReportVerdict) so no report ships
-          // verdict-less and lists as a permanent "?".
-          verdict: summary.verdict ? toVerdict(summary.verdict) : deriveReportVerdict(cases),
-          verifyRunId: runId,
-        });
-
-        // A case with no matching plan item means the run checked something it
-        // never planned — worth saying out loud, but not a failure. Only
-        // meaningful against a plan that actually names something: with no plan
-        // (or a cleared one) every case is trivially "unplanned", which is noise.
-        const unplanned = plan?.length
-          ? [...seenCheckItemIds].filter((id) => !plan.some((item) => item.id === id))
-          : [];
-
-        if (options.json !== undefined) {
-          outputJson(
-            {
-              acceptanceId,
-              cases: cases.length,
-              evidence: evidenceCount,
-              inlined,
-              origin,
-              planItems: plan?.length ?? 0,
-              pullRequest,
-              scenario,
-              subject: subject.ref,
-              unplanned,
-              verifyRunId: runId,
-            },
-            typeof options.json === 'string' ? options.json : undefined,
-          );
-          return;
-        }
-
-        console.log(
-          `${pc.green('✓')} Ingested ${pc.bold(String(cases.length))} case(s), ${pc.bold(String(evidenceCount))} evidence artifact(s)` +
-            `${inlined > 0 ? `, ${pc.bold(String(inlined))} inline` : ''}`,
-        );
-        if (plan?.length) {
-          const unexecuted = plan.filter((item) => !seenCheckItemIds.has(item.id));
-          console.log(
-            `${pc.bold('plan')}: ${plan.length} item(s)` +
-              `${unexecuted.length > 0 ? pc.yellow(` — ${unexecuted.length} planned but not executed`) : ''}` +
-              `${unplanned.length > 0 ? pc.dim(` — ${unplanned.length} unplanned case(s)`) : ''}`,
-          );
-        }
-        if (pullRequest?.url) console.log(`${pc.bold('pr')}: ${pullRequest.url}`);
-        if (origin?.topicId) console.log(`${pc.bold('origin topic')}: ${origin.topicId}`);
-        console.log(`${pc.bold('verifyRunId')}: ${runId} ${pc.dim('(immutable snapshot)')}`);
-        console.log(
-          `${pc.bold('acceptance')}: ${acceptanceId} ${pc.dim(`(${subject.ref.subjectType}:${subject.ref.subjectId})`)}`,
-        );
-        if (options.open) {
-          console.log(`${pc.bold('open')}: /verify/${runId}`);
-          console.log(`${pc.bold('open acceptance')}: /acceptance/${acceptanceId}`);
-        }
-      },
-    );
-}
-
-function printResults(results: any[]): void {
-  printTable(
-    results.map((r) => [
-      truncate(r.checkItemTitle || r.checkItemId, 50),
-      statusColor(r.status),
-      r.verdict ?? '',
-      r.confidence != null ? String(r.confidence) : '',
-      r.required ? 'gate' : 'soft',
-      truncate(r.suggestion || '', 40),
-    ]),
-    ['CHECK', 'STATUS', 'VERDICT', 'CONF', 'BLOCK', 'SUGGESTION'],
-  );
-}
-
-function statusColor(status: string): string {
-  if (status === 'passed') return pc.green(status);
-  if (status === 'failed') return pc.red(status);
-  if (status === 'running') return pc.yellow(status);
-  return pc.dim(status);
 }
